requirements.txt
# Core dependencies for ContextualAgentRulesHub
PyYAML>=6.0
